pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"


[project]
name = "pachyderm"
authors = [
  { name = "Raymond Ehlers", email = "raymond.ehlers@cern.ch" },
]
description = "Physics Analysis Core for Heavy-Ions"
readme = "README.md"
license.file = "LICENSE"
requires-python = ">=3.10"
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Science/Research",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: BSD License",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering",
  "Typing :: Typed",
]
dynamic = ["version"]
dependencies = [
    "attrs >=21.2.0",
    "numpy >=1.21",
    "scipy >=1.7.0",
    "matplotlib >=3.3.2",
    "iminuit >=2.7.0",
    "numdifftools >=0.9.40",
    "ruamel.yaml >=0.17.10",
]

[project.optional-dependencies]
test = [
    "pytest >=6",
    "pytest-cov >=3",
    "pytest-mock >=3.6.1",
    "pytest-xdist >=3.5.0",
    "pytest-sugar >=1.0.0",
]
dev = [
    "pre-commit >=2.13.0",
    "ruff >=0.0.209",
    "mypy >=0.931",
    "pytest >=6",
    "pytest-cov >=3",
    "pytest-mock >=3.6.1",
    "pytest-xdist >=3.5.0",
    "pytest-sugar >=1.0.0",
    "uproot >=4",
    "boost-histogram >=1.2.1",
    "hist >=2.7.1",
]
docs = [
    "sphinx>=7.0",
    "myst_parser>=0.13",
    "sphinx_copybutton",
    "sphinx_autodoc_typehints",
    "furo>=2023.08.17",
]

[project.urls]
Homepage = "https://github.com/raymondEhlers/pachyderm"
"Bug Tracker" = "https://github.com/raymondEhlers/pachyderm/issues"
Discussions = "https://github.com/raymondEhlers/pachyderm/discussions"
Changelog = "https://github.com/raymondEhlers/pachyderm/releases"

[project.scripts]
downloadALICEDataset = "pachyderm.alice.download:run_dataset_download"
downloadALICERunByRun = "pachyderm.alice.download:run_download_run_by_run_train_output"


[tool.hatch]
version.source = "vcs"
build.hooks.vcs.version-file = "src/pachyderm/_version.py"
build.include = [
    "src/pachyderm/alice/datasets/*.yaml",
]

[tool.hatch.envs.default]
features = ["test"]
scripts.test = "pytest {args}"


[tool.pytest.ini_options]
minversion = "6.0"
addopts = ["-ra", "--showlocals", "--strict-markers", "--strict-config"]
xfail_strict = true
filterwarnings = [
  "error",
]
log_level = "INFO"
log_cli_level = "INFO"
testpaths = [
  "tests",
]


[tool.coverage]
run.source = ["pachyderm"]
report.exclude_also = [
  '\.\.\.',
  'if typing.TYPE_CHECKING:',
]

[tool.mypy]
files = ["src", "tests"]
python_version = "3.10"
warn_unused_configs = true
strict = true
enable_error_code = ["ignore-without-code", "redundant-expr", "truthy-bool"]
warn_unreachable = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[[tool.mypy.overrides]]
module = "pachyderm.*"
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = ["ROOT", "numba", "nox", "iminuit", "numdifftools", "scipy.*"]
ignore_missing_imports = true


[tool.ruff]
src = ["src"]
exclude = ["src/pachyderm/_version.py"]  # Automatically generated, so there's no point in linting it
line-length = 120

[tool.ruff.lint]
extend-select = [
  "B",        # flake8-bugbear
  "I",        # isort
  "ARG",      # flake8-unused-arguments
  "C4",       # flake8-comprehensions
  "EM",       # flake8-errmsg
  "ICN",      # flake8-import-conventions
  "G",        # flake8-logging-format
  "PGH",      # pygrep-hooks
  "PIE",      # flake8-pie
  "PL",       # pylint
  "PT",       # flake8-pytest-style
  "PTH",      # flake8-use-pathlib
  "RET",      # flake8-return
  "RUF",      # Ruff-specific
  "SIM",      # flake8-simplify
  "T20",      # flake8-print
  "UP",       # pyupgrade
  "YTT",      # flake8-2020
  "EXE",      # flake8-executable
  "NPY",      # NumPy specific rules
]
ignore = [
  "PLR09",    # Too many <...>
  "PLR2004",  # Magic value used in comparison
  "ISC001",   # Conflicts with formatter
  "G004",     # Logging format string. This isn't best practices, but I use it too much to change for now
  "PLC0414",  # Messes with the re-export of imports needed from mypy
]
unfixable = [
    "T20",    # Removes print statements
    "F841",   # Removes unused variables
]

[tool.ruff.lint.mccabe]
max-complexity = 12

[tool.ruff.lint.isort]
required-imports = ["from __future__ import annotations"]
known-first-party = ["mammoth", "mammoth_cpp"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["T20"]
"noxfile.py" = ["T20"]


[tool.pylint]
py-version = "3.10"
ignore-paths = [".*/_version.py"]
reports.output-format = "colorized"
similarities.ignore-imports = "yes"
messages_control.disable = [
  "design",
  "fixme",
  "line-too-long",
  "missing-module-docstring",
  "wrong-import-position",
  "logging-fstring-interpolation",  # See above. I use it too much
]


[tool.codespell]
ignore-words-list = "thn, nd"
